outputs are correct now, with appropriate etas

+++ Code/Lib/lrgGradientDescentSolverStrategy.h
@@ -4,8 +4,8 @@ class GradientSolver : public DataSolverI {
     private:
     float eta; 
     int iterations;
-    int batch_size;
+    float batch_size;
     public:
-    GradientSolver(float eta,int iterations,int batch_size);
+    GradientSolver(float eta,int iterations,float batch_size);
     std::pair<double, double> &FitData(std::vector<std::pair<double, double> >&Inputs,std::pair<double, double>&Theta);
 };
